Experiment with ways to display javascript tasks
Tried to display code, test code, and results in a kind of grid: test code on top, other two on a row below.

It does work but it is tedious. Need to find a way to put the information in a json file and then read and display the file.

But is this possible without destroying code formatting?