created code to create next and previous buttons on child pages
Created code to create next and previous 'buttons' at the bottom of pages that are children of a main menu item.

This does not rely on 'order' being set.

Moved the code to a file in _include so it can be included where appropriate.

Created childpage.njk as a layout for pages that are children  of main menu items. This includes navnextprev.njk