created code to extract list of animals from data
The button are now generated automatically from a list of animals created from the actual data.

This means that I could replace the animal data and it will still work.

Could also make it generic so you can have a page of whatever objects you like as long as the data has the correct format.